+++ app/themes/salon/date-select-modal.html.php
@@ -1,0 +1,15 @@
+<!-- Date Select Modal -->
+<section id="date-select-modal" class="modal">
+  <div class="modal-inner">
+    <header>
+      <h3>Date Select</h3>
+      <a class="modal-close" onclick="F1.Modal.close(this, event)">Close X</a>
+    </header>
+    <div class="modal-body"><span id="date-nav-calendar">Calendar goes here...</span></div>
+    <footer>
+      <div class="actions-bar">
+        <div class="actions-left"></div>
+      </div>
+    </footer>
+  </div>
+</section>