+++ app/content/bookings/changelog.md
@@ -25,3 +25,23 @@
  - Add "Click on time-slot" to add a new booking.
  - Clean-up Edit Modal initialization code.
  - Reduce font size slightly in "Date Nav" and "New Button"
+
+## 22 Nov 2022 - Ver 2.2.0
+ - SAVE, EDIT and DELETE using AJAX! No more slow page reloads.
+ - Fix getting and setting the page state between page reloads. No more `date resets` on reload.
+ - Improve DB Service: insertOrUpdate(). Now return the ids of newly inserted records!
+ - Add HTTP service functions: getUrlParam() and getPostValue().
+ - Add ajax POST request support.
+ - Add delete button
+
+## 24 Nov 2022 - Ver 2.3.0
+ - Beautify booking summary and detail views - DONE
+ - Create a user with a lesser role, e.g. admin and validate against that to prevent unauthorized updates - DONE
+ - MRP - Click on date in Date Nav to open a calendar modal to select the current date - DONE
+ - Add an event handler for selecting the date from the calendar modal - DONE
+ - Fix issue with unwanted REQUEST params making SAVE fail! - DONE
+ - Make top part of calendar view sticky! Rrrr... - DONE
+ - Server side validation and feedback system - DONE
+ - MRP - Check if time-slot is booked - Validate - DONE
+ - MRP - Only creator can edit. Role system - DONE
+ - Use the current user's actual ID as created_by an updated_by - DONE